Solo falta comentarios

+++ src/unite1/Exercise04.java
@@ -9,12 +9,12 @@ public class Exercise04 {
 		Scanner ch = new Scanner(System.in);
 		
 		// Creamos la variable nota1.
-		int nota1;
+		double nota1;
 		
 		// Creamos la variable nota2.
-		int nota2;
+		double nota2;
 		
-		int media;
+		double media;
 		
 		System.out.println("¿Cual es tu primera nota?");
 		nota1 = ch.nextInt();
@@ -22,7 +22,7 @@ public class Exercise04 {
 		System.out.println("¿Cual es tu segunda nota?");
 		nota2 = ch.nextInt();
 		
-		media = (nota1 + nota2)/2;
+		media = (nota1 + nota2)/2.0;
 		
 		System.out.println("Esta es tu nota media: " + media);
